Reconfiguring build system to use cookoff
This [cookoff](https://github.com/UCL/GreatCMakeCookOff).
It should automatically download boost and eigen, if those libraries are
not found.
The next step is to make sure testing can be carried out in the build
tree, without installing.
And the last step would be to add an appropriate setup.py